+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - Add button for opening mobile data settings screen
   - Play ringtone when displaying the bundle balance toast message
   - Vibrate when showing bundle balance toast
+  - Handle bonga data bundles as lasting bundles
 
 + Misc